+++ app/app.py
@@ -1,21 +1,17 @@
-import asyncio
-import random
 from fastapi import FastAPI
-from fastapi import Request
-from fastapi.encoders import jsonable_encoder
-from fastapi.exceptions import HTTPException
-from fastapi.exceptions import RequestValidationError
 from fastapi.middleware.cors import CORSMiddleware
-from fastapi.responses import JSONResponse
-from app.routes import user
 from fastapi_pagination import add_pagination
+from fastapi.exceptions import HTTPException, RequestValidationError
 from app.middlewares.rate_limiter_middleware import RateLimitMiddleware
 from app.middlewares.request_id_injection import RequestIdInjection
-from pybreaker import CircuitBreakerError
-from dependencies import circuit_breaker
-from app.utils.slack_notification_utils import send_slack_message
-import traceback
-from app.middlewares.request_id_injection import request_id_contextvar
+
+from app.routes import api_router
+from app.utils.exception_handler import (
+    exception_handler,
+    validation_exception_handler,
+    http_exception_handler
+)
+
 
 # Initializing the swagger docs
 app = FastAPI(
@@ -23,8 +19,8 @@ app = FastAPI(
     description="This is my first API use FastAPI",
     version="0.0.1",
     openapi_tags=[{"name": "FastAPI Template", "description": "API template using FastAPI."}],
+    docs_url="/",
 )
-
 
 origins = ["*"]
 
@@ -38,75 +34,14 @@ app.add_middleware(
 )
 app.add_middleware(RateLimitMiddleware)
 app.add_middleware(RequestIdInjection)
-app.include_router(user, prefix="/user")
 
+# Include the routers
+app.include_router(api_router, prefix="/api")
 
-# Default API route
-@app.get("/")
-async def read_main():
-    print('Request ID:', request_id_contextvar.get())
-    return {"response": "service up and running..!"}
+# Exception handlers
+app.add_exception_handler(RequestValidationError, validation_exception_handler)
+app.add_exception_handler(HTTPException, http_exception_handler)
+app.add_exception_handler(Exception, exception_handler)
 
-
-async def external_service_call():
-    # Simulate network delay
-    delay = random.uniform(0.1, 1.0)  # Random delay between 0.1 to 1.0 seconds
-    await asyncio.sleep(delay)
-
-    # Simulate occasional failures
-    if random.random() < 0.2:  # 20% chance of failure
-        raise Exception("External service failed")
-
-    return "Success from external service"
-
-
-@app.get("/external-service")
-async def external_service_endpoint():
-    try:
-        with circuit_breaker:
-            result = await external_service_call()
-            return {"message": result}
-    except CircuitBreakerError:
-        raise HTTPException(status_code=503, detail="Service temporarily unavailable")
-    except Exception as e:
-        # Handle other exceptions from the external service call
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-# pylint: disable=unused-argument
-@app.exception_handler(RequestValidationError)
-async def validation_exception_handler(request: Request, exc: RequestValidationError):
-    return JSONResponse(
-        status_code=400,
-        content=jsonable_encoder({"message": "Validation error", "detail": exc.errors()[0]["msg"]}),
-    )
-
-
-# pylint: disable=unused-argument
-@app.exception_handler(HTTPException)
-async def http_exception_handler(request: Request, exc: HTTPException):
-    return JSONResponse(status_code=exc.status_code, content={"success": False, "message": exc.detail})
-
-@app.exception_handler(Exception)
-async def http_exception_handler(request: Request, exc: Exception):
-    error_message = f'Error: {str(exc)}'
-    # Include the traceback in the response for debugging purposes
-    traceback_str = traceback.format_exc(chain=False)
-    send_slack_message(
-        { 
-            "text": f'```\nRequestID: {request_id_contextvar.get()}\nRequest URL: {str(request.url)} \nRequest_method: {str(request.method)} \nTraceback: {traceback_str}```'
-        }
-    )
-    
-    return JSONResponse(
-        status_code=500,
-        content={"success": False, "message": error_message }
-    )
-
-
-@app.get("/{path:path}")
-async def catch_all(path: str):
-    return JSONResponse(status_code=404, content={"success": False, "message": f"Route not found for path: {path}"})
-
-
+# Add pagination support
 add_pagination(app)
