+++ app/routes/__init__.py
@@ -1,3 +1,11 @@
-from .users import user
+from fastapi import APIRouter
 
-__all__ = ["user"]
+from .home import home_router
+from .users import user_router
+
+
+api_router = APIRouter()
+api_router.include_router(user_router, prefix="/user")
+api_router.include_router(home_router, prefix="/home")
+
+__all__ = ["api_router"]
