+++ app/routes/home/__init__.py
@@ -1,0 +1,3 @@
+from .home import home_router
+
+__all__ = ["home_router"]
