+++ app/routes/users/__init__.py
@@ -1,0 +1,3 @@
+from .users import user_router
+
+__all__ = ["user_router"]
