+++ app/routes/users/users.py
@@ -18,26 +18,26 @@ from typing import Annotated
 from fastapi.security import HTTPBearer
 from app.middlewares.request_id_injection import request_id_contextvar
 
-user = APIRouter()
+user_router = APIRouter()
 
 httpBearerScheme = HTTPBearer()
 
 
-@user.post("/register", tags=["Users"])
+@user_router.post("/register", tags=["Users"])
 def register(payload: CreateUser, db: Session = Depends(create_local_session)):
     print('Request ID:', request_id_contextvar.get())
     response = create_user_dao(data=payload, dbSession=db)
     return response
 
 
-@user.post("/signin", tags=["Users"])
+@user_router.post("/signin", tags=["Users"])
 def login(payload: Login, db: Session = Depends(create_local_session)):
     print('Request ID:', request_id_contextvar.get())
     response = signin(data=payload, dbSession=db)
     return response
 
 
-@user.get("/{user_id}", tags=["Users"], dependencies=[Depends(get_current_user)])
+@user_router.get("/{user_id}", tags=["Users"], dependencies=[Depends(get_current_user)])
 async def profile(
     token: Annotated[str, Depends(httpBearerScheme)],
     user_id,
@@ -50,14 +50,14 @@ async def profile(
     return response
 
 
-@user.get("/", tags=["Users"], response_model=Page[UserOutResponse])
+@user_router.get("/", tags=["Users"], response_model=Page[UserOutResponse])
 def list_users(db: Session = Depends(create_local_session)):
     print('Request ID:', request_id_contextvar.get())
     response = list_users_dao(dbSession=db)
     return response
 
 
-@user.get("/{user_id}/secure-route/", tags=["Users"], dependencies=[Depends(get_current_user)])
+@user_router.get("/{user_id}/secure-route/", tags=["Users"], dependencies=[Depends(get_current_user)])
 def secure_route(token: Annotated[str, Depends(httpBearerScheme)], user_id: int):
     print('Request ID:', request_id_contextvar.get())
     return {"message": "If you see this, you're authenticated"}
